implemented export capability for HTML from markdown
