+++ apps/ds-ui/package.json
@@ -9,6 +9,7 @@
     "@uiw/react-md-editor": "^3.9.1",
     "bootstrap": "^5.1.3",
     "formik": "^2.2.9",
+    "js-file-download": "^0.4.12",
     "lodash-es": "^4.17.21",
     "luxon": "^2.4.0",
     "react": "^17.0.2",
