+++ packages/remark-plugins/package.json
@@ -6,6 +6,11 @@
   "types": "./build/index.d.ts",
   "dependencies": {
     "@devouringscripture/common": "^1.0.0",
+    "rehype-stringify": "^9.0.3",
+    "remark-gfm": "^3.0.1",
+    "remark-parse": "^10.0.1",
+    "remark-rehype": "^10.1.0",
+    "remark-supersub": "^1.0.0",
     "retext": "^8.1.0",
     "retext-smartypants": "^5.1.0",
     "unified": "^10.1.2",
