+++ packages/remark-plugins/src/index.ts
@@ -7,3 +7,4 @@ export * from './smartquotes/index';
 export * from './bible-poetry/index';
 export * from './all-cap-replacements/index';
 export * from './ad-bc-replacements/index';
+export * from './GetHTML/index';
